Fix isinstance check for score type to float

+++ Report Card Printer.py
@@ -7,4 +7,4 @@ print(is_student, type(is_student))
 age = 20
 print(age, type(age))
 score = 80.5
-print(isinstance(score, int))
+print(isinstance(score, float))
